requirements.txt
streamlit
requests
mistralai
python-dotenv